+++ app/styles/app.css
@@ -51,7 +51,7 @@ body {
     margin: 15px;
 }
 .coll-item {
-    height: 400px;
+    height: 200px;
     overflow: auto;
 }
 .coll-add-item {
@@ -79,4 +79,11 @@ body {
 }
 .coll-item-type-person {
     background-color: #8d98d8;
+}
+
+.coll-fake-mfr {
+    border: 1px solid #eee;
+    width: 100%;
+    height: 800px;
+    background: #808080;
 }
