+++ app/templates/collection/item.hbs
@@ -1,2 +1,24 @@
+{{bread-crumbs tagName="ol" outputStyle="bootstrap" linkable=true}}
+
 {{outlet}}
-This is item
+
+<div class="coll-item-page">
+    <div class="container">
+        <div class="row">
+            <div class="col-sm-4">
+                <h2> {{model.title}}</h2>
+                <p> {{model.description}}</p>
+                <div>
+                    {{#each model.tags as |tag|}}
+                        <span class="coll-item-tag">{{tag}}</span>
+                    {{/each}}
+                </div>
+            </div>
+            <div class="col-sm-8">
+                <div class="coll-mock-mfr">
+
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
