+++ package.json
@@ -35,6 +35,7 @@
     "ember-cli-test-loader": "^1.1.0",
     "ember-cli-uglify": "^1.2.0",
     "ember-collection": "1.0.0-alpha.7",
+    "ember-crumbly": "1.0.3",
     "ember-data": "^2.8.0",
     "ember-export-application-global": "^1.0.5",
     "ember-faker": "1.1.1",
